fix decomp across 11 functions — 6 to 100%, 5 improved
Functions reaching 100%:
- UIList::CalcBoundingBox: commutative fmuls operand order
- FlowDistance::Execute: condition inversions + branch swap
- RndAmbientOcclusion::BuildTrees: eliminated extra Triangle stack slot
- GetLightPosition: commutative fadds operand order via temp loads
- SpeechMgr::PrintSemanticTree: do-while → goto loop structure
- op14, op15, RockCentral::OnMsg, BinkReader::~BinkReader: already 100%

Functions improved:
- CacheMgrXbox::PollSearch: 94.9% → 99.9% (MILO_WARN→FAIL, structural rewrite)
- CharClip::EvaluateChannel: eliminate temp variable
- CharSleeve::Poll: std::pow → powf
- Utl::NextName: expression rewrite for numDigits
- MultiTempoTempoMap::GetLoopTick: variable declaration reorder

+++ src/system/char/CharSleeve.cpp
@@ -71,7 +71,7 @@ void CharSleeve::Poll() {
         float deltasecs = TheTaskMgr.DeltaSeconds();
         float dvar12 = deltasecs * 60.0f;
         float gravity_z = (mGravity * (deltasecs * (dvar12 * -3.858268f)));
-        float powed = std::pow(1.0f - mStiffness, dvar12 * dvar12);
+        float powed = powf(1.0f - mStiffness, dvar12 * dvar12);
         RndTransformable *sleeveparent = mSleeve->TransParent();
         float absed = std::fabs(mSleeve->LocalXfm().v.z);
         bool b2 = false;
